pyproject.toml
[project]
name = "subtitle-agent"
version = "0.1.0"
description = "A CLI toolset to find and download subtitles from SubDL and search TMDB."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "patool>=4.0.4",
    "pydantic>=2.12.5",
    "chardet>=5.2.0",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "pyinputplus>=0.2.12",
    "colorama>=0.4.6",
    "pygments>=2.19.2",
    "loguru>=0.7.3",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["src/cli/*"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 88
exclude = ["src/cli/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]
exclude = ["src/cli/*"]
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"
reportMissingTypeArgument = "none"
reportDeprecated = "none"
reportArgumentType = "none"
reportAttributeAccessIssue = "none"
